refactor: test files are no split and clarified depending on the scenario tested IHS-205
