--- tests/unit/doc_generation/mdx/test_mdx_ordered_code_doc.py
@@ -1,516 +1,0 @@
-"""Tests for OrderedMdxCodeDocumentation content reordering."""
-
-from __future__ import annotations
-
-import re
-from pathlib import Path
-from typing import TYPE_CHECKING
-from unittest.mock import MagicMock
-
-import pytest
-
-from docs.docs_generation.content_gen_methods.mdx.mdx_code_doc import ACodeDocumentation, MdxFile
-from docs.docs_generation.content_gen_methods.mdx.mdx_ordered_code_doc import OrderedMdxCodeDocumentation
-from docs.docs_generation.content_gen_methods.mdx.mdx_priority import PagePriority
-
-if TYPE_CHECKING:
-    from invoke import Context
-
-FILE_KEY = "test.mdx"
-MOCK_CONTEXT = MagicMock(spec="Context")
-MODULES: list[str] = []
-
-
-class TestReorderSections:
-    def test_section_priority_moves_classes_before_functions(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(sections=["Classes"])
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        assert _section_order(result) == ["Classes", "Functions"]
-
-    def test_combined_section_class_and_method_reordering(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(
-            sections=["Classes"],
-            classes=["InfrahubClient"],
-            methods={"InfrahubClient": ["save"]},
-        )
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        assert _section_order(result) == ["Classes", "Functions"]
-        assert _class_order(result)[0] == "InfrahubClient"
-        assert _method_order(result, "InfrahubClient")[0] == "save"
-
-    def test_no_section_priority_retains_original_order(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(sections=[])
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        assert result == sample_mdx
-
-
-class TestReorderClasses:
-    def test_single_priority_class_moves_to_top(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(classes=["InfrahubClient"])
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        order = _class_order(result)
-        assert order[0] == "InfrahubClient"
-
-    def test_multiple_priority_classes_in_specified_order(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(classes=["InfrahubClientSync", "InfrahubClient"])
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        order = _class_order(result)
-        assert order[0] == "InfrahubClientSync"
-        assert order[1] == "InfrahubClient"
-
-    def test_non_priority_classes_retain_original_order(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(classes=["InfrahubClient"])
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        order = _class_order(result)
-        remaining = order[1:]
-        assert remaining == ["ProcessRelationsNode", "BaseClient", "InfrahubClientSync"]
-
-    def test_no_priority_config_returns_unchanged(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority()
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        assert result == sample_mdx
-
-    def test_empty_classes_list_returns_unchanged(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(classes=[])
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        assert result == sample_mdx
-
-    def test_nonexistent_class_name_raises(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(classes=["DoesNotExist"])
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act / Assert
-        with pytest.raises(ValueError, match="DoesNotExist"):
-            doc.generate(MOCK_CONTEXT, MODULES)
-
-    def test_reorder_page_without_methods(self, sample_mdx_no_methods: str) -> None:
-        # Arrange
-        priority = PagePriority(classes=["InfrahubNodeMode"])
-        doc = _build_ordered_doc(sample_mdx_no_methods, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        order = _class_order(result)
-        assert order == ["InfrahubNodeMode", "RelatedNodeState"]
-
-
-class TestReorderMethods:
-    def test_single_priority_method_moves_to_top(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(methods={"InfrahubClient": ["save"]})
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        assert _method_order(result, "InfrahubClient")[0] == "save"
-
-    def test_multiple_priority_methods_in_specified_order(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(methods={"InfrahubClient": ["delete", "save"]})
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        order = _method_order(result, "InfrahubClient")
-        assert order[0] == "delete"
-        assert order[1] == "save"
-
-    def test_non_priority_methods_retain_original_order(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(methods={"InfrahubClient": ["save"]})
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        order = _method_order(result, "InfrahubClient")
-        assert order[0] == "save"
-        assert order[1:] == ["get_version", "create", "get", "get", "delete"]
-
-    def test_method_only_priority_no_class_reordering(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(methods={"InfrahubClient": ["save"]})
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        assert _class_order(result) == ["ProcessRelationsNode", "BaseClient", "InfrahubClient", "InfrahubClientSync"]
-
-    def test_combined_class_and_method_reordering(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(
-            classes=["InfrahubClient"],
-            methods={"InfrahubClient": ["save"]},
-        )
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        assert _class_order(result)[0] == "InfrahubClient"
-        assert _method_order(result, "InfrahubClient")[0] == "save"
-
-    def test_overloaded_methods_move_together(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(methods={"InfrahubClient": ["get"]})
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)[FILE_KEY].content
-
-        # Assert
-        order = _method_order(result, "InfrahubClient")
-        assert order[0] == "get"
-        assert order[1] == "get"
-        assert order[2:] == ["get_version", "create", "save", "delete"]
-
-    def test_nonexistent_method_name_raises(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(methods={"InfrahubClient": ["nonexistent", "save"]})
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act / Assert
-        with pytest.raises(ValueError, match="nonexistent"):
-            doc.generate(MOCK_CONTEXT, MODULES)
-
-    def test_method_priority_for_nonexistent_class_raises(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(methods={"DoesNotExist": ["get"]})
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act / Assert
-        with pytest.raises(ValueError, match="DoesNotExist"):
-            doc.generate(MOCK_CONTEXT, MODULES)
-
-
-class TestNoMatchingPriority:
-    def test_file_without_priority_returned_unchanged(self) -> None:
-        # Arrange
-        content = "# unchanged"
-        inner = _StubDocumentation({FILE_KEY: MdxFile(name=FILE_KEY, content=content, source_path=Path("test.py"))})
-        doc = OrderedMdxCodeDocumentation(documentation=inner, page_priorities={})
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)
-
-        # Assert
-        assert result[FILE_KEY].content == content
-
-
-class TestValidatePriorities:
-    def test_nonexistent_file_key_raises(self) -> None:
-        # Arrange
-        content = "# some content"
-        inner = _StubDocumentation(
-            {"actual.mdx": MdxFile(name="actual.mdx", content=content, source_path=Path("a.py"))}
-        )
-        doc = OrderedMdxCodeDocumentation(
-            documentation=inner,
-            page_priorities={"missing.mdx": PagePriority(classes=["Foo"])},
-        )
-
-        # Act / Assert
-        with pytest.raises(ValueError, match=r"missing\.mdx"):
-            doc.generate(MOCK_CONTEXT, MODULES)
-
-    def test_nonexistent_class_raises(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(classes=["NoSuchClass"])
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act / Assert
-        with pytest.raises(ValueError, match="NoSuchClass"):
-            doc.generate(MOCK_CONTEXT, MODULES)
-
-    def test_nonexistent_method_raises(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(methods={"InfrahubClient": ["no_such_method"]})
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act / Assert
-        with pytest.raises(ValueError, match="no_such_method"):
-            doc.generate(MOCK_CONTEXT, MODULES)
-
-    def test_valid_config_no_error(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(
-            classes=["InfrahubClient"],
-            methods={"InfrahubClient": ["save"]},
-        )
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act
-        result = doc.generate(MOCK_CONTEXT, MODULES)
-
-        # Assert
-        assert FILE_KEY in result
-
-    def test_duplicate_class_names_raises(self) -> None:
-        # Act / Assert
-        with pytest.raises(ValueError, match="Duplicate class 'InfrahubClient'"):
-            PagePriority(classes=["InfrahubClient", "InfrahubClient"])
-
-    def test_duplicate_method_names_raises(self) -> None:
-        # Act / Assert
-        with pytest.raises(ValueError, match="Duplicate method 'save'"):
-            PagePriority(methods={"InfrahubClient": ["save", "save"]})
-
-    def test_duplicate_section_names_raises(self) -> None:
-        # Act / Assert
-        with pytest.raises(ValueError, match="Duplicate section 'Classes'"):
-            PagePriority(sections=["Classes", "Classes"])
-
-    def test_nonexistent_section_raises(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(sections=["NoSuchSection"])
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act / Assert
-        with pytest.raises(ValueError, match="NoSuchSection"):
-            doc.generate(MOCK_CONTEXT, MODULES)
-
-    def test_method_class_not_in_file_raises(self, sample_mdx: str) -> None:
-        # Arrange
-        priority = PagePriority(methods={"GhostClass": ["get"]})
-        doc = _build_ordered_doc(sample_mdx, priority)
-
-        # Act / Assert
-        with pytest.raises(ValueError, match="GhostClass"):
-            doc.generate(MOCK_CONTEXT, MODULES)
-
-
-# --- Helpers ---
-
-
-class _StubDocumentation(ACodeDocumentation):
-    """Minimal stub returning pre-built MdxFile dicts."""
-
-    def __init__(self, files: dict[str, MdxFile]) -> None:
-        self._files = files
-
-    def generate(self, context: Context, modules_to_document: list[str]) -> dict[str, MdxFile]:
-        return self._files
-
-
-def _build_ordered_doc(content: str, priority: PagePriority) -> OrderedMdxCodeDocumentation:
-    """Build an ``OrderedMdxCodeDocumentation`` with a stub inner documentation."""
-    inner = _StubDocumentation({FILE_KEY: MdxFile(name=FILE_KEY, content=content, source_path=Path("test.py"))})
-    return OrderedMdxCodeDocumentation(documentation=inner, page_priorities={FILE_KEY: priority})
-
-
-def _section_order(content: str) -> list[str]:
-    """Extract the order of H2 section names."""
-    return re.findall(r"^## (\w+)", content, re.MULTILINE)
-
-
-def _class_order(content: str) -> list[str]:
-    """Extract the order of H3 class names under ``## Classes``."""
-    match = re.search(r"^## Classes\n(.*?)(?=^## |\Z)", content, re.MULTILINE | re.DOTALL)
-    if not match:
-        return []
-    return re.findall(r"^### `([^`]+)`", match.group(1), re.MULTILINE)
-
-
-def _method_order(content: str, class_name: str) -> list[str]:
-    """Extract the order of H4 method names under a given H3 class section."""
-    pattern = rf"^### `{re.escape(class_name)}`\n(.*?)(?=^### |\Z)"
-    match = re.search(pattern, content, re.MULTILINE | re.DOTALL)
-    if not match:
-        return []
-    return re.findall(r"^#### `([^`]+)`", match.group(1), re.MULTILINE)
-
-
-# --- Fixtures ---
-
-
-@pytest.fixture
-def sample_mdx() -> str:
-    return """\
----
-title: client
-sidebarTitle: client
----
-
-# `infrahub_sdk.client`
-
-## Functions
-
-### `handle_relogin`
-
-```python
-handle_relogin(func: Callable) -> Callable
-```
-
-### `handle_relogin_sync`
-
-```python
-handle_relogin_sync(func: Callable) -> Callable
-```
-
-## Classes
-
-### `ProcessRelationsNode`
-
-Process relations for a node.
-
-### `BaseClient`
-
-Base class for InfrahubClient and InfrahubClientSync
-
-**Methods:**
-
-#### `start_tracking`
-
-```python
-start_tracking(self) -> Self
-```
-
-#### `set_context_properties`
-
-```python
-set_context_properties(self, identifier: str) -> None
-```
-
-### `InfrahubClient`
-
-GraphQL Client to interact with Infrahub.
-
-**Methods:**
-
-#### `get_version`
-
-```python
-get_version(self) -> str
-```
-
-Return the Infrahub version.
-
-#### `create`
-
-```python
-create(self, kind: str) -> InfrahubNode
-```
-
-#### `get`
-
-```python
-get(self, kind: str) -> InfrahubNode
-```
-
-#### `get`
-
-```python
-get(self, kind: type[SchemaType]) -> SchemaType
-```
-
-#### `save`
-
-```python
-save(self, node: InfrahubNode) -> None
-```
-
-#### `delete`
-
-```python
-delete(self, kind: str, id: str) -> None
-```
-
-### `InfrahubClientSync`
-
-Synchronous GraphQL Client to interact with Infrahub.
-
-**Methods:**
-
-#### `get_version`
-
-```python
-get_version(self) -> str
-```
-
-#### `create`
-
-```python
-create(self, kind: str) -> InfrahubNodeSync
-```
-"""
-
-
-@pytest.fixture
-def sample_mdx_no_methods() -> str:
-    return """\
----
-title: constants
-sidebarTitle: constants
----
-
-# `infrahub_sdk.node.constants`
-
-## Classes
-
-### `RelatedNodeState`
-
-State of a related node.
-
-### `InfrahubNodeMode`
-
-Mode of an Infrahub node.
-"""
